requirements.txt
gdown
tensorflow==2.7.0
transformers
gitpython
wget
torch
scikit-learn
progressbar2
gensim==3.8.3
sentence-transformers
tensorflow-hub==0.12.0
torch==1.10.1
numpy==1.19
pandas